@@ -16,21 +16,22 @@
     "@aws-sdk/client-textract": "^3.5.0",
     "@aws-sdk/lib-storage": "^3.5.0",
     "axios": "^0.21.1",
+    "bootstrap": "^4.6.0",
     "cheerio": "^1.0.0-rc.5",
+    "datatables.net": "^1.10.23",
+    "datatables.net-bs4": "^1.10.23",
     "debug": "^4.3.1",
+    "jquery": "^3.5.1",
+    "popper.js": "^1.16.1",
     "street-suffix": "^1.0.1",
     "url-to-s3": "^1.0.0",
     "uuid": "^8.3.2",
     "yargs": "^16.2.0",
     "core-js": "^3.6.5",
-    "vue": "^2.6.11"
+    "vue": "^3.0.0",
+    "vue-router": "^4.0.3"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-eslint": "~4.5.0",
-    "@vue/cli-service": "~4.5.0",
-    "eslint-plugin-vue": "^6.2.2",
-    "vue-template-compiler": "^2.6.11",
     "@types/node": "^14.14.9",
     "eslint": "^7.16.0",
     "eslint-config-prettier": "^6.15.0",
@@ -38,7 +39,12 @@
     "mocha": "^8.3.0",
     "nodemon": "^2.0.7",
     "prettier": "^2.2.0",
-    "babel-eslint": "^10.1.0"
+    "@vue/cli-plugin-babel": "~4.5.0",
+    "@vue/cli-plugin-eslint": "~4.5.0",
+    "@vue/cli-service": "~4.5.0",
+    "@vue/compiler-sfc": "^3.0.0-0",
+    "babel-eslint": "^10.1.0",
+    "eslint-plugin-vue": "^7.0.0-0"
   },
   "eslintConfig": {
     "root": true,
@@ -52,7 +58,7 @@
     ],
     "extends": [
       "eslint:recommended",
-      "plugin:vue/essential",
+      "plugin:vue/vue3-essential",
       "prettier"
     ],
     "parserOptions": {
