@@ -1,8 +1,8 @@
-import Vue from 'vue';
+import { createApp } from 'vue';
 import App from './App.vue';
 
-Vue.config.productionTip = false;
+import 'bootstrap'; //js code
+import 'bootstrap/dist/css/bootstrap.min.css';
 
-new Vue({
-  render: (h) => h(App)
-}).$mount('#app');
+const app = createApp(App);
+app.mount('#app');
