@@ -1,6 +1,6 @@
 <template>
   <main>
-    {{ message }}
+    <h1>{{ message }} <span class="badge badge-secondary">New</span></h1>
   </main>
 </template>
 
@@ -14,14 +14,3 @@ export default {
   }
 };
 </script>
-
-<style>
-#app {
-  font-family: Avenir, Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  text-align: center;
-  color: #2c3e50;
-  margin-top: 60px;
-}
-</style>
